refactor(pages): tela de informacao do usuario no contexto de recuperacao de senha movida de main.dart para pasta 'pages'
